refactor directory of option pricing
pricing models and strike calculation moved under strategies/calculators.
