+++ src/trading_strategies/strategy/option_strategy/calculators/option_strike.py
