+++ src/trading_strategies/strategy/option_strategy/long_put.py
@@ -4,9 +4,8 @@ from src.market.order import Order
 from src.trading_strategies.financial_asset.option import PutOption
 from src.trading_strategies.financial_asset.price import Price, EmptyPrice
 from src.trading_strategies.financial_asset.symbol import Symbol
-from src.trading_strategies.option_pricing import bsm_pricing
 from src.trading_strategies.strategy.option_strategy.option_strategy import OptionStrategy
-from src.trading_strategies.strategy.option_strategy.option_strike import calculate_strike
+from src.trading_strategies.strategy.option_strategy.calculators.option_strike import calculate_strike
 from src.trading_strategies.strategy.strategy_id import StrategyId
 from src.agent.transactions.position import Position
 from src.util.expiry_date import next_expiry_date
