+++ test/test_option_pricing.py
@@ -4,7 +4,7 @@ from src.trading_strategies.financial_asset.option import CallOption, PutOption
 from src.trading_strategies.financial_asset.stock import Stock
 from src.trading_strategies.financial_asset.price import Price
 from src.trading_strategies.financial_asset.symbol import Symbol
-from src.trading_strategies.option_pricing import bsm_pricing
+from src.trading_strategies.strategy.option_strategy.calculators.option_pricing import bsm_pricing
 
 
 class TestOptionPricing(unittest.TestCase):
